adapt disposition to other signup methods
