+++ ephios/event_management/signup.py
@@ -21,6 +21,7 @@ from django.views import View
 
 from ephios.event_management.models import AbstractParticipation, LocalParticipation, Shift
 from ephios.extra.widgets import CustomSplitDateTimeWidget
+from ephios.plugins.basesignup.signup.disposition import BaseDispositionParticipationForm
 from ephios.user_management.models import Qualification
 
 register_signup_methods = django.dispatch.Signal()
@@ -217,6 +218,13 @@ class BaseSignupMethod:
     signup_error_message = _("Signing up failed: {error}")
     decline_success_message = _("You have successfully declined {shift}.")
     decline_error_message = _("Declining failed: {error}")
+
+    """
+    This form will be used for participations in disposition.
+    Set to None if you don't want to support the default disposition.
+    """
+    disposition_participation_form_class = BaseDispositionParticipationForm
+    disposition_show_requested_state = True
 
     def __init__(self, shift):
         self.shift = shift
