+++ ephios/plugins/basesignup/signals.py
@@ -1,11 +1,9 @@
 from django.dispatch import receiver
 
 from ephios.event_management.signup import register_signup_methods
+from ephios.plugins.basesignup.signup.instant import InstantConfirmationSignupMethod
+from ephios.plugins.basesignup.signup.request_confirm import RequestConfirmSignupMethod
 from ephios.plugins.basesignup.signup.section_based import SectionBasedSignupMethod
-from ephios.plugins.basesignup.signup.simple import (
-    InstantConfirmationSignupMethod,
-    RequestConfirmSignupMethod,
-)
 
 
 @receiver(
