+++ ephios/plugins/basesignup/signup/section_based.py
@@ -6,26 +6,22 @@ from operator import itemgetter
 from django import forms
 from django.contrib import messages
 from django.core.exceptions import ValidationError
-from django.http import Http404
 from django.shortcuts import redirect
 from django.template.loader import get_template
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
-from django.views import View
-from django.views.generic import FormView, TemplateView
-from django.views.generic.base import TemplateResponseMixin
-from django.views.generic.detail import SingleObjectMixin
-from django_select2.forms import ModelSelect2Widget, Select2MultipleWidget
+from django.views.generic import FormView
+from django_select2.forms import Select2MultipleWidget
 
-from ephios.event_management.models import AbstractParticipation, Shift
+from ephios.event_management.models import AbstractParticipation
 from ephios.event_management.signup import (
     AbstractParticipant,
     BaseSignupMethod,
     BaseSignupView,
     ParticipationError,
 )
-from ephios.extra.permissions import CustomPermissionRequiredMixin
-from ephios.user_management.models import Qualification, UserProfile
+from ephios.plugins.basesignup.signup.disposition import BaseDispositionParticipationForm
+from ephios.user_management.models import Qualification
 
 
 def sections_participant_qualifies_for(sections, participant: AbstractParticipant):
@@ -37,7 +33,9 @@ def sections_participant_qualifies_for(sections, participant: AbstractParticipan
     ]
 
 
-class DispositionParticipationForm(forms.ModelForm):
+class SectionBasedDispositionParticipationForm(BaseDispositionParticipationForm):
+    disposition_participation_template = "basesignup/section_based/fragment_participant.html"
+
     section = forms.ChoiceField(
         label=_("Section"),
         required=False,  # only required if participation is confirmed
@@ -46,18 +44,9 @@ class DispositionParticipationForm(forms.ModelForm):
         ),
     )
 
-    class Meta:
-        model = AbstractParticipation
-        fields = ["state", "id"]
-        widgets = dict(state=forms.HiddenInput(attrs={"class": "state-input"}))
-
-    def __init__(self, shift, **kwargs):
+    def __init__(self, **kwargs):
         super().__init__(**kwargs)
-
-        # if "id" not in self.fields:
-        #    self.fields['id'] = forms.IntegerField(initial=self.instance.pk, widget=forms.HiddenInput)
-
-        sections = shift.signup_method.configuration.sections
+        sections = self.shift.signup_method.configuration.sections
         qualified_sections = list(
             sections_participant_qualifies_for(
                 sections,
@@ -104,123 +93,6 @@ class DispositionParticipationForm(forms.ModelForm):
     def save(self, commit=True):
         self.instance.data["dispatched_section_uuid"] = self.cleaned_data["section"]
         super().save(commit)
-
-
-DispositionParticipationFormset = forms.modelformset_factory(
-    model=AbstractParticipation,
-    form=DispositionParticipationForm,
-    extra=0,
-    can_order=False,
-    can_delete=True,
-)
-
-
-def addable_users(shift):
-    """
-    Return queryset of user objects that can be added to the shift.
-    This also includes users that already have a participation, as that might have gotten removed in JS.
-
-    This also includes users that can normally not see the event. The permission will be added accordingly.
-    # TODO: check how that looks and wether we actually check for that elsewhere and also actually add the permission.
-    """
-    return UserProfile.objects.all()  # you surprised it's just this? :D
-
-
-class AddUserForm(forms.Form):
-    user = forms.ModelChoiceField(
-        widget=ModelSelect2Widget(
-            model=UserProfile,
-            search_fields=["first_name__icontains", "last_name__icontains"],
-            attrs={"form": "add-user-form"},
-        ),
-        queryset=UserProfile.objects.none(),  # set using __init__
-    )
-
-    def __init__(self, user_queryset, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["user"].queryset = user_queryset
-
-
-class SectionBasedAddUserView(
-    CustomPermissionRequiredMixin, SingleObjectMixin, TemplateResponseMixin, View
-):
-    template_name = "basesignup/section_based/fragment_participant.html"
-    permission_required = "event_management.change_event"
-    model = Shift
-
-    def setup(self, request, *args, **kwargs):
-        super().setup(request, *args, **kwargs)
-        self.object: Shift = self.get_object()
-
-    def get_permission_object(self):
-        return self.object.event
-
-    def post(self, request, *args, **kwargs):
-        shift = self.object
-        form = AddUserForm(
-            user_queryset=addable_users(shift),
-            data=request.POST,
-        )
-        if form.is_valid():
-            user: UserProfile = form.cleaned_data["user"]
-            instance = shift.signup_method.get_participation_for(user.as_participant())
-            instance.state = AbstractParticipation.States.RESPONSIBLE_ADDED
-            instance.save()
-
-            formset = DispositionParticipationFormset(
-                queryset=self.object.participations,
-                prefix="participations",
-                form_kwargs=dict(shift=self.object),
-            )
-            form = next(filter(lambda form: form.instance.id == instance.id, formset))
-            return self.render_to_response({"form": form})
-        raise Http404("User does not exist")
-
-
-class SectionBasedDispositionView(CustomPermissionRequiredMixin, SingleObjectMixin, TemplateView):
-    # TODO unite this class with the request_confirm one ;)  [and the other ones too]
-    model = Shift
-    permission_required = "event_management.change_event"
-    template_name = "basesignup/disposition.html"
-
-    def setup(self, request, *args, **kwargs):
-        super().setup(request, *args, **kwargs)
-        self.object: Shift = self.get_object()
-
-    def get_permission_object(self):
-        return self.object.event
-
-    def get_formset(self):
-        formset = DispositionParticipationFormset(
-            self.request.POST or None,
-            queryset=self.object.participations,
-            prefix="participations",
-            form_kwargs=dict(shift=self.object),
-        )
-        return formset
-
-    def post(self, request, *args, **kwargs):
-        formset = self.get_formset()
-        if formset.is_valid():
-            formset.save()
-            self.object.participations.filter(
-                state=AbstractParticipation.States.RESPONSIBLE_ADDED
-            ).delete()
-            return redirect(self.object.event.get_absolute_url())
-        return self.get(request, *args, **kwargs, formset=formset)
-
-    def get_context_data(self, **kwargs):
-        kwargs.setdefault("formset", self.get_formset())
-        kwargs.setdefault("states", AbstractParticipation.States)
-        kwargs.setdefault("sections", self.object.signup_method.configuration.sections)
-        kwargs.setdefault(
-            "participant_template", "basesignup/section_based/fragment_participant.html"
-        )
-        kwargs.setdefault(
-            "add_user_form",
-            AddUserForm(user_queryset=addable_users(self.object)),
-        )
-        return super().get_context_data(**kwargs)
 
 
 class SectionForm(forms.Form):
@@ -333,8 +205,11 @@ class SectionBasedSignupMethod(BaseSignupMethod):
     registration_button_text = _("Request")
     signup_success_message = _("You have successfully requested a participation for {shift}.")
     signup_error_message = _("Requesting a participation failed: {error}")
+
     configuration_form_class = SectionBasedConfigurationForm
     signup_view_class = SectionBasedSignupView
+
+    disposition_participation_form_class = SectionBasedDispositionParticipationForm
 
     def get_configuration_fields(self):
         return {
@@ -415,7 +290,7 @@ class SectionBasedSignupMethod(BaseSignupMethod):
                 "confirmed_sections_with_users": confirmed_sections_with_users,
                 "disposition_url": (
                     reverse(
-                        "basesignup:shift_disposition_section_based",
+                        "basesignup:shift_disposition",
                         kwargs=dict(pk=self.shift.pk),
                     )
                     if request.user.has_perm("event_management.change_event", obj=self.shift.event)
