+++ ephios/plugins/basesignup/templates/basesignup/common/disposition.html
@@ -19,7 +19,7 @@
     </div>
 
     <form method="post" class="d-none" id="add-user-form"
-          action="{% url "basesignup:shift_disposition_add_user_section_based" pk=shift.pk %}">{% csrf_token %}</form>
+          action="{% url "basesignup:shift_disposition_add_user" pk=shift.pk %}">{% csrf_token %}</form>
 
     <form method="post" class="form" id="participation-form" data-formset data-formset-prefix="{{ formset.prefix }}">
         {% csrf_token %}
@@ -61,19 +61,21 @@
                 </div>
                 <div class="col-lg-6 d-flex flex-column">
 
-                    <div class="card border-warning mb-3 flex-grow-1">
-                        <div class="card-body d-flex flex-column">
-                            <h5 class="card-title">{{ states.REQUESTED.label|capfirst }}</h5>
-                            <div class="list-group list-group-flush flex-grow-1"
-                                 data-drop-to-state="{{ states.REQUESTED }}">
-                                {% for form in formset %}
-                                    {% if form.instance.state == states.REQUESTED %}
-                                        {% include participant_template %}
-                                    {% endif %}
-                                {% endfor %}
+                    {% if shift.signup_method.disposition_show_requested_state %}
+                        <div class="card border-warning mb-3 flex-grow-1">
+                            <div class="card-body d-flex flex-column">
+                                <h5 class="card-title">{{ states.REQUESTED.label|capfirst }}</h5>
+                                <div class="list-group list-group-flush flex-grow-1"
+                                     data-drop-to-state="{{ states.REQUESTED }}">
+                                    {% for form in formset %}
+                                        {% if form.instance.state == states.REQUESTED %}
+                                            {% include participant_template %}
+                                        {% endif %}
+                                    {% endfor %}
+                                </div>
                             </div>
                         </div>
-                    </div>
+                    {% endif %}
 
                     <div class="card border-danger mb-3 flex-grow-1">
                         <div class="card-body d-flex flex-column">
