+++ ephios/plugins/basesignup/templates/basesignup/common/fragment_participant.html
@@ -1,0 +1,20 @@
+{% load bootstrap4 %}
+{% load i18n %}
+<div class="list-group-item draggable" data-formset-form data-participant-id="{{ form.instance.user.id }}">
+    <h6 class="mb-0">
+        <div class="row align-items-center no-gutters">
+            <div class="col-10 pr-3">
+                <span class="fa fa-user"></span>
+                {{ form.instance.participant }}
+            </div>
+            <div class="col-2 d-flex justify-content-between align-items-center">
+                <button type="button" class="btn float-right btn-sm ml-3" data-formset-delete-button>
+                    <i class="fas fa-times"></i>
+                </button>
+            </div>
+        </div>
+    </h6>
+    <div class="d-none">
+        {% bootstrap_form form layout="inline" %}
+    </div>
+</div>
