+++ ephios/plugins/basesignup/templates/basesignup/instant/fragment_state.html
@@ -11,3 +11,9 @@
         </ul>
     </dd>
 </dl>
+
+{% if disposition_url %}
+<div class="row float-right">
+    <a href="{{ disposition_url }}" class="btn btn-outline-info"><span class="fas fa-hand-point-right"></span> {% translate "Disposition" %}</a>
+</div>
+{% endif %}
