--- ephios/plugins/basesignup/templates/basesignup/requestconfirm/fragment_participant.html
@@ -1,10 +1,0 @@
-{% load bootstrap4 %}
-<li class="list-group-item draggable">
-    <h6 class="card-title mb-0">
-        <span class="fa fa-user"></span>
-        {{ form.instance.participant }}
-    </h6>
-    <div class="d-none">
-        {{ form }}
-    </div>
-</li>
